🤖 Refactor(client): mockup 생성 코드 이동
